--- src/main/java/com/rshu/mppz/CustomerDetails.java
@@ -1,5 +1,0 @@
-package com.rshu.mppz;
-
-public class CustomerDetails {
-
-}
