+++ src/main/java/com/rshu/mppz/customer/data/CustomerRepository.java
@@ -1,4 +1,4 @@
-package com.rshu.mppz;
+package com.rshu.mppz.customer.data;
 
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.Page;
