+++ src/main/java/com/rshu/mppz/order/controller/OrderController.java
@@ -1,5 +1,12 @@
 
-package com.rshu.mppz;
+package com.rshu.mppz.order.controller;
+
+import com.rshu.mppz.product.controller.Product;
+import com.rshu.mppz.product.controller.ProductRepository;
+import com.rshu.mppz.customer.data.Customer;
+import com.rshu.mppz.customer.data.CustomerRepository;
+import com.rshu.mppz.order.data.Order;
+import com.rshu.mppz.order.data.OrderRepository;
 
 import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
